Clean up code for online publication pt1
All the following were done for experiment 1 only:

Remove analysis from the codes, so now the codes only include codes to read the model outputs straight from the rds file in folder.

Analyses were checked and ensured they matched the paper, 

Remove Bayesian analyses for version difference 

Remove ordered restricted model analyses 

some problems encountered with setwd(), not sure why but It seem to set directory automatically to the location of the script. The error encountered was:

"The working directory was changed to C:/Intel inside a notebook chunk. The working directory will be reset when the chunk is finished running. Use the knitr root.dir option in the setup chunk to change the working directory for notebook chunks."